+++ src/components/command/index.ts
@@ -1,2 +1,1 @@
 export * from "./command-palette"
-export * from "./advanced-command-menu"