+++ src/index.ts
@@ -25,7 +25,6 @@ export * from './components/ui/spinner'
 export * from './components/ui/tooltip'
 export * from './components/ui/kbd'
 export * from './components/ui/radio-group'
-export * from './components/ui/scroll-box'
 export * from './components/ui/right-click-menu'
 export * from './components/ui/toggle-group'